requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlalchemy==2.0.49
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
cryptography==47.0.0
pydantic==2.13.3
pydantic-settings==2.14.0
httpx==0.28.1
python-multipart==0.0.26
aiofiles==25.1.0
email-validator==2.2.0
